third_party/blink/web_tests/external/wpt/css/css-multicol/parsing/column-pseudo-computed-expected.txt

This is a testharness.js-based test.
[FAIL] ::column scroll-snap-margin
  assert_equals: expected (string) "2px" but got (undefined) undefined
Harness: the test ran to completion.
